# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for the crowdfunding projects service.",
        "title": "Project Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:10001",
    "basePath": "/",
    "paths": {
        "/projects": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "List all projects",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Project"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "Create project",
                "parameters": [
                    {
                        "description": "Project data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateProjectReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Project"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/projects/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "Get project by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Project"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Category": {
            "type": "string",
            "enum": [
                "science",
                "tech",
                "architecture_and_urban",
                "sport",
                "music"
            ],
            "x-enum-varnames": [
                "CategoryScience",
                "CategoryTech",
                "CategoryArchitectureAndUrban",
                "CategorySport",
                "CategoryMusic"
            ]
        },
        "domain.CreateProjectReq": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/domain.Category"
                },
                "currency": {
                    "$ref": "#/definitions/domain.Currency"
                },
                "description": {
                    "type": "string"
                },
                "duration_days": {
                    "type": "integer"
                },
                "goal_amount": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "domain.Currency": {
            "type": "string",
            "enum": [
                "RUB",
                "USD"
            ],
            "x-enum-varnames": [
                "CurrencyRUB",
                "CurrencyUSD"
            ]
        },
        "domain.Project": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/domain.Category"
                },
                "currency": {
                    "$ref": "#/definitions/domain.Currency"
                },
                "description": {
                    "type": "string"
                },
                "duration_days": {
                    "type": "integer"
                },
                "goal_amount": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.Status"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "domain.Status": {
            "type": "string",
            "enum": [
                "draft",
                "review",
                "active",
                "finished"
            ],
            "x-enum-varnames": [
                "StatusDraft",
                "StatusReview",
                "StatusActive",
                "StatusFinished"
            ]
        }
    }
}